+++ lib/src/proxy/upload-data.ts
@@ -1,200 +1,16 @@
 // Copyright 2026 The Swarm Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { Reference } from "@ethersphere/bee-js"
-import type {
-  Bee,
-  BeeRequestOptions,
-  Stamper,
-  UploadOptions,
-  UploadResult,
-} from "@ethersphere/bee-js"
-import {
-  makeContentAddressedChunk,
-  type ContentAddressedChunk,
-  DEFAULT_UPLOAD_CONCURRENCY,
-} from "../chunk"
-import { splitDataIntoChunks, buildMerkleTree } from "./chunking"
-import type { UploadContext, UploadProgress, ChunkReference } from "./types"
-import { tryCreateTag } from "../utils/tag"
-
 /**
- * Simple Uint8ArrayWriter implementation for ChunkAdapter
+ * Plain data upload functions.
+ *
+ * @deprecated This module is deprecated. For new code, use the unified API:
+ *   import { uploadData, uploadChunk } from "./upload"
+ *
+ * This module re-exports legacy wrappers for backward compatibility.
  */
-class SimpleUint8ArrayWriter {
-  cursor: number = 0
-  buffer: Uint8Array
 
-  constructor(buffer: Uint8Array) {
-    this.buffer = buffer
-  }
-
-  write(_reader: unknown): number {
-    throw new Error("SimpleUint8ArrayWriter.write() not implemented")
-  }
-
-  max(): number {
-    return this.buffer.length
-  }
-}
-
-/**
- * Adapter to convert bee-js Chunk to cafe-utility Chunk interface
- * This allows the Stamper to work with content-addressed chunks
- */
-class ChunkAdapter {
-  span: bigint
-  writer: SimpleUint8ArrayWriter
-
-  constructor(private chunk: ContentAddressedChunk) {
-    this.span = chunk.span.toBigInt()
-    this.writer = new SimpleUint8ArrayWriter(chunk.data)
-  }
-
-  hash(): Uint8Array {
-    return this.chunk.address.toUint8Array()
-  }
-
-  build(): Uint8Array {
-    return this.chunk.data
-  }
-}
-
-/**
- * Upload data with client-side signing
- * Handles chunking, merkle tree building, and progress reporting
- */
-export async function uploadDataWithSigning(
-  context: UploadContext,
-  data: Uint8Array,
-  options?: UploadOptions,
-  onProgress?: (progress: UploadProgress) => void,
-  requestOptions?: BeeRequestOptions,
-): Promise<{ reference: string; tagUid?: number }> {
-  const { bee, stamper } = context
-
-  // Create a tag for tracking upload progress (required for fast deferred uploads)
-  const tag = options?.tag ?? (await tryCreateTag(bee))
-
-  // Step 1: Split data into chunks
-  const chunkPayloads = splitDataIntoChunks(data)
-  let totalChunks = chunkPayloads.length
-  let processedChunks = 0
-
-  // Progress callback helper
-  const reportProgress = () => {
-    if (onProgress) {
-      onProgress({ total: totalChunks, processed: processedChunks })
-    }
-  }
-
-  // Step 2: Process leaf chunks
-  const chunkRefs: ChunkReference[] = []
-
-  // Merge tag into options for all chunk uploads
-  const uploadOptionsWithTag = { ...options, tag }
-
-  // Create all chunks first (CPU work)
-  const chunks: ContentAddressedChunk[] = []
-  for (const payload of chunkPayloads) {
-    const chunk = makeContentAddressedChunk(payload)
-    chunks.push(chunk)
-    chunkRefs.push({ address: chunk.address.toUint8Array() })
-  }
-
-  // Upload in parallel batches
-  for (let i = 0; i < chunks.length; i += DEFAULT_UPLOAD_CONCURRENCY) {
-    const batch = chunks.slice(
-      i,
-      Math.min(i + DEFAULT_UPLOAD_CONCURRENCY, chunks.length),
-    )
-    await Promise.all(
-      batch.map(async (chunk) => {
-        await uploadSingleChunk(
-          bee,
-          stamper,
-          chunk,
-          uploadOptionsWithTag,
-          requestOptions,
-        )
-        processedChunks++
-        reportProgress()
-      }),
-    )
-  }
-
-  // Step 3: Build merkle tree (if multiple chunks)
-  let rootReference: Reference
-
-  if (chunkRefs.length === 1) {
-    // Single chunk - use direct reference
-    rootReference = new Reference(chunkRefs[0].address)
-  } else {
-    // Multiple chunks - build tree
-
-    rootReference = await buildMerkleTree(
-      chunkRefs,
-      async (intermediateChunk) => {
-        await uploadSingleChunk(
-          bee,
-          stamper,
-          intermediateChunk,
-          uploadOptionsWithTag,
-          requestOptions,
-        )
-
-        // Count intermediate chunks in progress
-        totalChunks++
-        processedChunks++
-        reportProgress()
-      },
-    )
-  }
-
-  // Return result
-  return {
-    reference: rootReference.toHex(),
-    tagUid: tag,
-  }
-}
-
-/**
- * Upload a single chunk with optional signing
- * Returns UploadResult with Bee's actual stored reference for verification
- */
-export async function uploadSingleChunk(
-  bee: Bee,
-  stamper: Stamper | undefined,
-  chunk: ContentAddressedChunk,
-  options?: UploadOptions,
-  requestOptions?: BeeRequestOptions,
-): Promise<UploadResult> {
-  // Use non-deferred mode for faster uploads (returns immediately)
-  // Note: pinning is incompatible with deferred mode, so disable it
-  const uploadOptions = { deferred: false, ...options }
-
-  if (stamper) {
-    // Client-side signing - use adapter for cafe-utility Chunk interface
-    const chunkAdapter = new ChunkAdapter(chunk)
-    const envelope = stamper.stamp(chunkAdapter)
-    const result = await bee.uploadChunk(
-      envelope,
-      chunk.data,
-      uploadOptions,
-      requestOptions,
-    )
-
-    // Verify address match between locally computed and Bee's response
-    const localAddress = chunk.address.toHex()
-    const beeAddress = result.reference.toHex()
-    if (localAddress !== beeAddress) {
-      console.error(
-        `[ADDRESS MISMATCH] local=${localAddress} bee=${beeAddress}`,
-      )
-    }
-
-    return result
-  } else {
-    throw new Error("No stamper or batch ID available")
-  }
-}
+export {
+  uploadDataWithSigning,
+  uploadSingleChunk,
+} from "./upload-legacy-wrappers"
